Move clickhouse_kafka_sender.py to src/base
